+++ src/FlabIt.Guardians/StringGuardiansExtension.cs
@@ -76,18 +76,20 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentEmptyException">Raised when <paramref name="argument"/> is empty.</exception>
+        /// <returns>The <paramref name="argument" /> when it is not null or empty.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfNullOrEmpty(
+        [NotNull]
+        public static string ThrowIfNullOrEmpty(
             [CanBeNull, ValidatedNotNull] this string? argument,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsEmpty(argument))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -107,12 +109,13 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfNullOrEmpty()' in upcoming versions.")]
         public static string PassThroughNonNullNorEmpty(
             [CanBeNull, ValidatedNotNull] this string? argument,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsEmpty(argument))
                 return argument;
@@ -133,16 +136,19 @@ namespace FlabIt.Guardians
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentEmptyException">Raised when <paramref name="argument"/> is empty.</exception>
         /// <exception cref="ArgumentWhitespaceException">Raised when <paramref name="argument"/> consists only of whitespaces.</exception>
+        /// <returns>The <paramref name="argument" /> when it is not null nor empty nor consists only of whitespaces.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfNullOrEmptyOrWhitespace(
+        [NotNull]
+        public static string ThrowIfNullOrEmptyOrWhitespace(
             [CanBeNull, ValidatedNotNull] this string? argument,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNullOrEmpty(argumentName, message);
-            argument.ThrowIfNullOrWhitespace(argumentName, message);
+            return argument
+                .ThrowIfNullOrEmpty(argumentName, message)
+                .ThrowIfNullOrWhitespace(argumentName, message);
         }
 
         /// <summary>
@@ -159,6 +165,7 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfNullOrEmptyOrWhitespace()' in upcoming versions.")]
         public static string PassThroughNonNullNorEmptyNorWhitespace(
             [CanBeNull, ValidatedNotNull] this string? argument,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
@@ -178,18 +185,20 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentWhitespaceException">Raised when <paramref name="argument"/> consists only of whitespaces.</exception>
+        /// <returns>The <paramref name="argument" /> when it is not null nor consists only of whitespaces.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfNullOrWhitespace(
+        [NotNull]
+        public static string ThrowIfNullOrWhitespace(
             [CanBeNull, ValidatedNotNull] this string? argument,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsOnlyWhitespaces(argument))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -209,12 +218,13 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfNullOrWhitespace()' in upcoming versions.")]
         public static string PassThroughNonNullNorWhitespace(
             [CanBeNull, ValidatedNotNull] this string? argument,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsOnlyWhitespaces(argument))
                 return argument;
@@ -234,19 +244,21 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentLengthShorterThanException">Raised when <paramref name="argument"/> is shorter than <paramref name="length"/>.</exception>
+        /// <returns>The <paramref name="argument" /> when it has the same or greater length than specified with <paramref name="length"/>.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfShorterThan(
-            [NotNull, ValidatedNotNull] this string? argument,
+        [NotNull]
+        public static string ThrowIfShorterThan(
+            [NotNull, ValidatedNotNull] this string argument,
             int length,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsShorterThan(argument, length))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -267,8 +279,9 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfShorterThan()' in upcoming versions.")]
         public static string PassThroughNonShorterThan(
-            [NotNull, ValidatedNotNull] this string? argument,
+            [NotNull, ValidatedNotNull] this string argument,
             int length,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
@@ -293,19 +306,21 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentLengthLargerThanException">Raised when <paramref name="argument"/> is larger than <paramref name="length"/>.</exception>
+        /// <returns>The <paramref name="argument" /> when it has the same or shorter length than specified with <paramref name="length"/>.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfLargerThan(
-            [NotNull, ValidatedNotNull] this string? argument,
+        [NotNull]
+        public static string ThrowIfLargerThan(
+            [NotNull, ValidatedNotNull] this string argument,
             int length,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsLargerThan(argument, length))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -326,8 +341,9 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfLargerThan()' in upcoming versions.")]
         public static string PassThroughNonLargerThan(
-            [NotNull, ValidatedNotNull] this string? argument,
+            [NotNull, ValidatedNotNull] this string argument,
             int length,
             [CanBeNull, InvokerParameterName] string? argumentName = null,
             [CanBeNull] string? message = null)