+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 
 * Set target framework to .Net 5.0
 * Updated NuGet packages
+* Marked PassThrough... guardians as obsolete as their ThrowIf... counterpart now is able to return validated input arguments.
 
 ## v0.6